// rusticasaFront/src/app/shared/model/requests/request-registrar-casa-.model.ts
export class RequestRegistrarCasa {
  idCasa?: number;
  descripcion: string;
  nombreCasa: string;
  mascotas: boolean;
  precioNoche: number;
  numeroHabitaciones: number;
  numeroInquilinos: number;
  piscina: boolean;
  wifi: boolean;
  jardin: boolean;
  idMunicipio: number;
  gmail: string;
}